+++ public/injected-config.js
@@ -1,0 +1,3 @@
+// Embroider expects this file to be here since it's being referenced in `index.html`.
+// Config injection only happens on staging and production environments, however.
+console.debug('DEBUG: Config injection is working!');